+++ services/accountmanager/dirk/service_internal_test.go
@@ -1,4 +1,4 @@
-// Copyright © 2020 Attestant Limited.
+// Copyright © 2020, 2021 Attestant Limited.
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
@@ -190,6 +190,7 @@ func setupService(ctx context.Context, t *testing.T, endpoints []string, account
 		WithLogLevel(zerolog.TraceLevel),
 		WithMonitor(nullmetrics.New(context.Background())),
 		WithClientMonitor(nullmetrics.New(context.Background())),
+		WithProcessConcurrency(1),
 		WithEndpoints(endpoints),
 		WithAccountPaths(accountPaths),
 		WithClientCert([]byte(resources.ClientTest01Crt)),